+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/EntityTurn.java
@@ -1,0 +1,6 @@
+package com.github.mjkuranda.spaceadventure2.entities;
+
+public enum EntityTurn {
+    INCOMING,
+    OUTCOMING
+}