+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/Moveable.java
@@ -1,0 +1,5 @@
+package com.github.mjkuranda.spaceadventure2.entities;
+
+public interface Moveable {
+    void move();
+}