+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/shootable/Shootable.java
@@ -1,0 +1,5 @@
+package com.github.mjkuranda.spaceadventure2.entities.shootable;
+
+public interface Shootable {
+    void shoot();
+}